Add various functions and tests
It may be useful to start using a Vec<Vec<_>>. This is because view_box works optimally under
those conditions. It it rare that we traverse along the y axis anyways, and even if we did,
there'd be no benefit in any spatial locatity considering the large distances on huge
tilenets. I'll need to change that. Then, iterators using skip can skip huge portions
of unnecessary parts.

Another method is to implement my own iterator, this would actually be preferable.
The iterator will thus also not show internal information.

It would work by simply referencing and counting where it is in the box.
Very simply, and will be efficient even for massive maps (where linear scanning
may take some time).