Add the collide_set function
This function returns a TileSet over all the tiles that we're in
collision with. This is useful if you can provide an iterator of
locations you'd like to check. For example an object's corners
can be encapsulated as an iterator, then you get a new iterator
containing all potential collisions.