Use the natural end number for f32 to denote the limit
16777216 is the maximum float number that can be represented with 100% accuracy.
The next number is 16777218. This means that the algorithm that is used will not function outside
this range. The range must therefore be limited.
I'll need to study the effects at the boundary of the range. One potential scare
is the addition of dx and dy. We know sx and sy are both
always -1 or 1, which means they'll be accurate to 16777216.
The dx won't. To keep precision maximal, maybe an
error accumulator needs to be used:

	if self.ey > 1.0 {
		self.ey -= self.sy;
		self.iy += self.sy;
	}

Or something like that. I'll need to experiment in matlab.