Add a bunch of tests
The removal of minimization is based on ex and ey being very different
in some cases. It then makes little sense to minimize one.
If ex is 0.001 and ey is 16000000, then subtracting will not have any effect on ey, so it's best to just say that the
algorithm is designed for floats smaller than a certain value.